test(sqlite): relocate integration tests to e2e + unit modules
The branch's engine_crud test duplicated the in-memory DB plumbing
that the rebased develop now provides via tests/e2e.rs. Move the
query-many decode case into e2e.rs using the existing
PRAX_E2E / SQLITE_URL harness.

The three row_ref.rs tests didn't need to be in tests/ at all —
they construct an rusqlite Connection synchronously and exercise
SqliteRowRef's pure-library behaviour. Move them into
src/row_ref.rs's #[cfg(test)] mod tests so they run under the
normal unit-test suite (and don't gate behind PRAX_E2E).
